+++ src/ast/mod.rs
@@ -1,7 +1,9 @@
 mod lexer;
+mod lexer_types;
 mod nodes;
 mod parser;
 
-pub use self::lexer::*;
 pub use self::nodes::*;
 pub use self::parser::*;
+pub use self::lexer_types::*;
+pub use self::lexer::*;
